Start IPCR

+++ YamangTao.Api/Dtos/Pms/RatingDto.cs
@@ -3,7 +3,7 @@ namespace YamangTao.Api.Dtos.Pms
     public class RatingDto
     {
         // public uint id { get; set; }
-        public uint RatingMatrixId { get; set; }
+        public long RatingMatrixId { get; set; }
         public sbyte Rate { get; set; }
         public string Description { get; set; }
     }
