+++ YamangTao.Model/PM/Rating.cs
@@ -3,7 +3,7 @@ namespace YamangTao.Model.PM
     public class Rating
     {
         // public uint id { get; set; }
-        public uint RatingMatrixId { get; set; }
+        public long RatingMatrixId { get; set; }
         public RatingMatrix Matrix { get; set; }
         public sbyte Rate { get; set; }
         public string Description { get; set; }
